langgraph/plan-and-execute/pyproject.toml
[project]
name = "langgraph-plan-and-execute"
version = "0.1.0"
description = "LangGraph plan-and-execute tutorial implementation"
requires-python = ">=3.9"
dependencies = [
    "langgraph",
    "langchain-openai",
    "langchain-core",
    "langchain-community",
    "python-dotenv",
    "langchain-tavily",
]

[tool.uv]
dev-dependencies = []
